Revert to using fabric.
