+++ setup.py
@@ -69,9 +69,9 @@ dev_requirements = [
     "selenium==2.44.0",
 
     # The cloud acceptance test runner needs these
+    "fabric==1.10.0",
     "apache-libcloud==0.16.0",
     "digitalocean-python==0.1.5",
-    "crochet==1.3.0",
 
     # Packages are downloaded from Buildbot
     "requests==2.4.3",
